Fix the view for Ingredient Entries
The accepted params did not include :ingredient so it was silently
failing in the controller. This PR fixes that, and also adds a separate
table for recently updated ingredient entries (updated on the current
day). This PR also addresses an issue whereby attributes on the
IngredientEntry that were empty strings, rather than nil, were not being
processed correctly.