Add basic authentication to controllers
This PR adds a basic check to the controllers to render a 404 not found
on all non-public-facing controllers if the client's IP address does not
equal my home IP address. This is an interim fix until proper
authentication can be implemented with Devise.

It also does some tidying up of unnecessary routes and renames the
IngredientsController as it had the wrong name.