Recipe JSON schema
Adds a method to Recipe that generates the standard JSON in the standard
schema required for rich search results in Google searches. The JSON has
been marked as html_safe to ensure it is rendered in the head
successfully.

Also includes some renaming of files and the addition of the
IngredientSet factory which was previously missing.