Add basic validation to the form using ActiveModel::Validations
This commit adds validation to the contact form on the menu page using a
combination of JS and regular rails Action View, due to the form being
unconventional in that it deals with a @contact_form_message instead of
an Order.

It was not possible to have the error messages from the model available
in the JS file. The controller re-renders the form if the
ContactFormMessage is invalid, passing the errors through as a local
variable and keeps the inputted values populated. A conditional in the
view them renders an error div if the invalid email error message is present. Errors
for the other fields are handled by the JS.