Add new admin panel to blog posts
This commit adds the new admin panel to the Blog Post show page. The
changes also include the new routes and actions to the controller, and
the updates to the BlogPostMailer to use images in the asset folder
instead of the s3 image path saved on the blog post record.